Modify classes (creation/deletion registrations)
Modify classes:
- ru.dreremin.predefense.registration.sys.controllers.create
	.CreateRegistrationController;
- ru.dreremin.predefense.registration.sys.controllers.delete
	.DeleteRegistrationController;
- ru.dreremin.predefense.registration.sys.services.registrations
	.CreateRegistrationService;
- ru.dreremin.predefense.registration.sys.services.registrations.
	DeleteRegistrationService.
Add classes:
- ru.dreremin.predefense.registration.sys.requestdto.LoginDto;
- ru.dreremin.predefense.registration.sys.exceptions.ExpiredComissionException.

+++ src/main/java/ru/dreremin/predefense/registration/sys/repositories/TeacherComissionRepository.java
@@ -16,7 +16,7 @@ public interface TeacherComissionRepository
 	
 	List<TeacherComission> findByTeacherId(long teacherId);
 	
-	Optional<TeacherComission> findByTeacherIdAndComissionId(long teacherId, 
+	Optional<TeacherComission> findByTeacherIdAndComissionId(int teacherId, 
 															 int comissionId);
 	
 	void deleteAllByTeacherId(int teacherId);
